requirements.txt
PySide6>=6.5.0
psycopg2-binary>=2.9.0
